requirements.txt
flask
matplotlib
seaborn
numpy
ultralytics
flask
flask-login
flask-sqlalchemy
pillow
lapx